Moved some reusable code from routes.py to controllers.py to prepare to unit test them
